Phase 5: rename Interactive Explorer to /explorer.html with redirects
Closes the unified-explorer migration (issue #156). Phases 1-4 (#157-#160)
built the unified UI on tutorials/progressive_globe.qmd; this PR promotes
it to the canonical site-root URL and retires the old Search Explorer
page.

Rename + asset-path fix
- Move tutorials/progressive_globe.qmd → explorer.qmd at site root.
- Adjust source-palette import from `../assets/js/source-palette.js` to
  `assets/js/source-palette.js` so it resolves on both isamples.org and
  on rdhyee.github.io/isamplesorg.github.io PR previews.

URL param: search, not q
- The explorer's URL state now uses `?search=` instead of `?q=`. Quarto's
  site-wide search reserves `?q=` for its highlight feature and strips
  the param via history.replaceState before any of our cells run (see
  docs/site_libs/quarto-search/quarto-search.js). `?search=` is unused
  by Quarto and survives intact.

Redirect stubs at the old URLs
- tutorials/progressive_globe.html and tutorials/isamples_explorer.html
  become preview-safe redirect stubs:
    new URL('../explorer.html' + search + hash, href)
- They forward whatever query string the browser presents. Note: legacy
  `?q=basalt` URLs lose the search term because Quarto strips `?q=`
  before our stub script runs (the stub is itself a Quarto-rendered
  page, so its <head> loads quarto-search.js). Non-q params (sources,
  material, etc.) and the hash fragment all survive — the only
  affected URLs are Phase 3 dev test links that were never published.

_quarto.yml + internal links
- Navbar Interactive Explorer href → explorer.qmd. Search Explorer
  removed from both the How-to-Use menu and the sidebar.
- Update internal links to /explorer.html in index.qmd, how-to-use.qmd,
  tutorials/index.qmd, about.qmd, data.qmd, design/index.qmd,
  index_alt.qmd, query-spec.qmd, tutorials/narrow_vs_wide_performance.qmd,
  and the existing parquet_cesium_isamples_wide redirect stub.

Tests
- tests/test_explorer.py → tests/test_globe.py targeting /explorer.html.
- Selectors updated for the unified DOM-based UI: #sourceFilter,
  #materialFilter, #contextFilter, #objectTypeFilter, #globeViewBtn /
  #tableViewBtn (no List view), #maxSamples number input.
- Unskip the cross-filter facet tests deferred in #155 — native HTML
  checkboxes respond to programmatic .click() unlike the old Explorer's
  OJS Inputs.checkbox.
- Add redirect-preserves-params tests for both old URLs (using the
  current ?search= param, which survives Quarto's q-stripping).
- test_navigation.py + test_tutorials_landing.py drop Search Explorer
  assertions and retarget the globe-loads test to /explorer.html.
