+++ _quarto.yml
@@ -12,14 +12,12 @@ website:
     left:
       - href: index.qmd
         text: Home
-      - href: tutorials/progressive_globe.qmd
+      - href: explorer.qmd
         text: Interactive Explorer
       - text: How to Use
         menu:
           - text: Overview
             href: how-to-use.qmd
-          - text: Search Explorer
-            href: tutorials/isamples_explorer.qmd
           - text: Deep-Dive Analysis
             href: tutorials/zenodo_isamples_analysis.qmd
       - text: About
@@ -67,8 +65,6 @@ website:
       contents:
         - href: how-to-use.qmd
           text: Overview
-        - text: "Search Explorer"
-          href: tutorials/isamples_explorer.qmd
         - text: "Deep-Dive Analysis"
           href: tutorials/zenodo_isamples_analysis.qmd
     - id: about
