+++ tutorials/narrow_vs_wide_performance.qmd
@@ -1068,6 +1068,6 @@ The wide format eliminates ~9M edge rows (60% file size reduction). The H3 varia
 
 ## See Also
 
-- [Interactive Explorer](/tutorials/progressive_globe.html) - Browse samples on a 3D globe
+- [Interactive Explorer](/explorer.html) - Browse samples on a 3D globe
 - [Search Explorer](isamples_explorer.qmd) - Faceted search and filter across all samples
 - [Deep-Dive Analysis](zenodo_isamples_analysis.qmd) - Comprehensive DuckDB-WASM tutorial
